Add functionality to add a new book to the library
Clicking on 'New Book' button will prompt user to enter information for the new book to add. If any of the inputs are empty, the form will not submit. The user also has the option to close the form without adding a new book by pressing the 'Close' button. Submitting a new book will populate in on the library display.